Housing prices and attributes from 1976.

Housing prices and attributes from 1976.

hprice

Format

A data frame with 506 rows and 12 variables, including:

lprice

log median house price in the community

nox

Nitrous Oxide (NOX) concentration; parts per million

crime

number of reported crimes per capita

rooms

average number of rooms in houses in the community

dist

weighted distance of the community to 5 employment centers

stratio

average student-teacher ratio of schools in the community

proptax

property taxes

...

Source

Harrison, Jr., David, Rubinfeld, Daniel L. (1978/03)."Hedonic housing prices and the demand for clean air." Journal of Environmental Economics and Management 5(1): 81-102. <http://hdl.handle.net/2027.42/22636> https://deepblue.lib.umich.edu/handle/2027.42/22636

Note

hprice is useful for teaching hedonic pricing (namely, the relative value of pollution as captured by housing prices), and linear regression in general.