Simulate date from a simple linear model

Simulate date from a simple linear model

simlm(
  model = NULL,
  N = 100,
  betas,
  error,
  seed = sample.int(.Machine$integer.max, 1),
  ...
)

Arguments

model
a model or formula, e.g., `y ~ x1 + x2`

N
number of observations. Defaults to 100

betas
coefficients of the right-hand side variables

error
distribution of the error term. Defaults to standard normal: `rnorm(n=N, mean = 0, sd = 1)`

seed
random number generator seed

...
distributions of the right-hand side variables (see demo below) using any of R's probability distribution functions (e.g., `rnorm`, `rbinom`, `rpois`, etc)

Value

data.frame

Note

Lots of work to do on this one (in particular: heteroskedasticity, instrumental variables, multicollinearity)

Author

Lawrence R. De Geest, lrdegeest@gmail.com

Examples

# generate random data where the RHS variables are normally distributed
df = simlm(model = y ~ x1 + x2,
           x1 = rnorm(n=N, mean = 10, sd = 3),
           x2 = rnorm(n=N, mean = 20, sd = 6),
           betas = c(x1 = 3, x2 = 5))
# "recover" the parameters with `lm`:
lm(y ~ x1 + x2, data = df)

#> 
#> Call:
#> lm(formula = y ~ x1 + x2, data = df)
#> 
#> Coefficients:
#> (Intercept)           x1           x2  
#>        1.73         2.95         4.99  
#> 

# demonstrate endogeneity in the form of omitted variable bias by excluding a term:
lm(y ~ x1, data = df)

#> 
#> Call:
#> lm(formula = y ~ x1, data = df)
#> 
#> Coefficients:
#> (Intercept)           x1  
#>      97.279        3.303  
#> 

# generate an "experiment" with one treatment/control condition and increase the noise:
experiment = simlm(model = response ~ treatment + x1 + x2,
           treatment = rbinom(n=N, size = 1, prob = 0.5),
           x1 = rnorm(n=N, mean = 20, sd = 6),
           x2 = rpois(n=N, lambda = 5),
           betas = c(treatment = -4, x1 = 2.4, x2 = 7.89)) # treatment has a negative effect